django_blog/requirements.txt
django>=5.0,<6.0
psycopg2-binary>=2.9
gunicorn>=21.2
python-dotenv>=1.0
Pillow>=10.0
markdown>=3.5
django-markdownify>=0.9
mysql-connector-python>=8.0